+++ CHANGES.rst
@@ -11,6 +11,11 @@ Breaking changes:
 
 New features:
 
+- Remove jquery.cookie from plone-logged-in bundle's stub_js_modules.
+  The toolbar, which has a dependency on jquery.cookie,
+  was moved from the plone bundle to plone-logged-in in CMPlone 5.1a2.
+  [thet]
+
 - Products.MimetypesRegistry has no longer a skins layer, remove it.
   [jensens]